Refactor out the 'active' field in stopwatchtimer struct
It is possible know the state simply by looking at the start/end time.

If start time is non-zero and end timer is zero, the timer is running.
If start time is zero or end time is non-zero the timer is not running.

So we remove the needless 'active' field.