Store the state as real time so it can be used across reboots
The monotonic clock is good for measuring time intervals but the start
time is undefined, so it is not usable across reboots.

We store the time as real (wall) time and convert it back to monotonic
at reload. This way it works across reboots.

Also increase the buffer for storing the time stamp. It needs to contain
micorseconds since 1 Jan 1970.

Change the type to gint64, which is what glib uses for
g_get_{monotonic,real}_time(). This allows us to have a negative start
time in case elapsed time is bigger than monotonic clocks start (elapsed
time is bigger than uptime?)